Admin: список товаров, редактирование, добавление, удаление.

+++ src/main/java/com/example/internship/service/impl/ProductServiceImpl.java
@@ -7,9 +7,12 @@ import com.example.internship.repository.ProductRepository;
 import com.example.internship.service.ProductService;
 import com.example.internship.specification.ProductSpecification;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
+
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -29,17 +32,22 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public List<ProductDto> findAll() {
-        return productRepo.findAll().stream().map(this::convertToDto).collect(Collectors.toList());
+        return productRepo.findAll(Sort.by("id")).stream().map(this::convertToDto).collect(Collectors.toList());
     }
 
     @Override
     public void removeProduct(Long id) {
-        productRepo.deleteById(id);
+        if (productRepo.existsById(id)) productRepo.deleteById(id);
     }
 
     @Override
     public void addProduct(ProductDto productDto) {
         productRepo.save(convertToModel(productDto));
+    }
+
+    @Override
+    public void saveProduct(Product product) {
+        productRepo.save(product);
     }
 
     @Override
@@ -50,9 +58,12 @@ public class ProductServiceImpl implements ProductService {
                 .collect(Collectors.toList());
     }
 
+    public Product findByIdProduct(Long id) {
+        return productRepo.findById(id).orElseThrow();
+    }
+
     @Override
     public ProductDto getProductById(Long id) {
-
         return convertToDto(productRepo.findById(id).get());
     }
 
@@ -101,4 +112,5 @@ public class ProductServiceImpl implements ProductService {
     private Product convertToModel(ProductDto productDto) {
         return mapper.map(productDto, Product.class);
     }
+
 }
